requirements.txt
einops==0.3.2
h5py==3.6.0
joblib==0.17.0
matplotlib==3.4.3
natsort==7.0.1
numpy==1.21.3
Pillow==9.1.0
pip==21.3.1
scikit_image==0.18.3
scipy==1.7.1
skimage==0.0
timm==0.4.12
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
